OpenAI compatible provider set up for vision
