+++ chibi/services/providers/moonshotai.py
@@ -8,6 +8,7 @@ class MoonshotAI(OpenAIFriendlyProvider):
     api_key = gpt_settings.moonshotai_key
     chat_ready = True
     moderation_ready = True
+    vision_ready = True
 
     base_url = "https://api.moonshot.ai/v1"
     name = "MoonshotAI"
@@ -17,4 +18,5 @@ class MoonshotAI(OpenAIFriendlyProvider):
     image_size = NOT_GIVEN
     default_model = "kimi-latest"
     default_moderation_model = "kimi-k2-turbo-preview"
+    default_vision_model = "kimi-k2.5"
     temperature = 1
