+++ chibi/services/providers/openai.py
@@ -1,4 +1,3 @@
-import base64
 from io import BytesIO
 
 from loguru import logger
@@ -7,8 +6,7 @@ from openai.types import ImagesResponse, ReasoningEffort
 
 from chibi.config import gpt_settings
 from chibi.constants import OPENAI_TTS_INSTRUCTIONS
-from chibi.schemas.app import VisionResultSchema
-from chibi.services.providers.provider import OpenAIFriendlyProvider, ServiceResponseError
+from chibi.services.providers.provider import OpenAIFriendlyProvider
 
 
 class OpenAI(OpenAIFriendlyProvider):
@@ -44,58 +42,6 @@ class OpenAI(OpenAIFriendlyProvider):
             logger.info(f"Transcribed text: {response.text}")
             return response.text
         raise ValueError("Could not transcribe audio message")
-
-    async def vision(
-        self,
-        image: bytes,
-        mime_type: str,
-        model: str | None = None,
-        prompt: str | None = None,
-    ) -> VisionResultSchema:
-        model = model or self.default_vision_model
-        prompt = prompt or "Describe the image in detail."
-        logger.info(f"Analyzing image with model {model}...")
-
-        # Encode image to base64
-        image_base64 = base64.b64encode(image).decode("utf-8")
-        data_url = f"data:{mime_type};base64,{image_base64}"
-
-        # Use parse() for structured output with Pydantic models
-        response = await self.get_client().chat.completions.parse(
-            model=model,
-            messages=[
-                {
-                    "role": "user",
-                    "content": [
-                        {"type": "text", "text": prompt},
-                        {
-                            "type": "image_url",
-                            "image_url": {"url": data_url, "detail": "high"},
-                        },
-                    ],
-                }
-            ],
-            response_format=VisionResultSchema,
-            max_tokens=4096,
-        )
-
-        if not response.choices:
-            raise ServiceResponseError(
-                provider=self.name,
-                model=model,
-                detail=f"Could not analyze image: empty response: {response}",
-            )
-
-        result = response.choices[0].message
-        if not result or not result.parsed:
-            raise ServiceResponseError(
-                provider=self.name,
-                model=model,
-                detail=f"Could not analyze image: empty response: {response}",
-            )
-
-        logger.info(f"Image analyzed successfully: {result.parsed.short_description[:50]}...")
-        return result.parsed
 
     async def speech(self, text: str, voice: str | None = None, model: str | None = None) -> bytes:
         voice = voice or self.default_tts_voice
