+++ test_file.txt
@@ -1,0 +1,3 @@
+start
+content
+end
